+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### 0.6.4
+
+- **Fix**: SpinnerView bug(no animation) in Swift 4.
+
+## 0.6.3
+- **Fix**: SeparatorCell updates constraints after setting separator insets
+
 ## 0.6.2
 - **Fix**: AlamofireManager extension no longer performs requests with default manager